+++ src/PharmacyAdjudicator/PharmacyAdjudicator.DataAccess/Rule.cs
@@ -20,7 +20,7 @@ namespace PharmacyAdjudicator.DataAccess
             this.PlanRules = new HashSet<PlanRules>();
         }
     
-        public int RuleId { get; set; }
+        public long RuleId { get; set; }
         public string RuleType { get; set; }
         public string DefaultValue { get; set; }
